refactor : alcohol 패키지로 이동 및 record로 변환
