+++ src/main/java/app/bottlenote/alcohols/repository/custom/CustomAlcoholQueryRepository.java
@@ -1,10 +1,10 @@
 package app.bottlenote.alcohols.repository.custom;
 
 import app.bottlenote.alcohols.dto.dsl.AlcoholSearchCriteria;
+import app.bottlenote.alcohols.dto.response.AlcoholInfo;
 import app.bottlenote.alcohols.dto.response.AlcoholSearchResponse;
 import app.bottlenote.alcohols.dto.response.detail.AlcoholDetailInfo;
 import app.bottlenote.global.service.cursor.PageResponse;
-import app.bottlenote.review.dto.response.AlcoholInfo;
 
 public interface CustomAlcoholQueryRepository {
