+++ src/main/java/app/bottlenote/alcohols/service/domain/AlcoholDomainSupport.java
@@ -1,7 +1,7 @@
 package app.bottlenote.alcohols.service.domain;
 
 import app.bottlenote.alcohols.domain.Alcohol;
-import app.bottlenote.review.dto.response.AlcoholInfo;
+import app.bottlenote.alcohols.dto.response.AlcoholInfo;
 import java.util.Optional;
 
 public interface AlcoholDomainSupport {
