+++ src/main/java/app/bottlenote/alcohols/service/domain/DefaultAlcoholDomainSupport.java
@@ -2,7 +2,7 @@ package app.bottlenote.alcohols.service.domain;
 
 import app.bottlenote.alcohols.domain.Alcohol;
 import app.bottlenote.alcohols.domain.AlcoholQueryRepository;
-import app.bottlenote.review.dto.response.AlcoholInfo;
+import app.bottlenote.alcohols.dto.response.AlcoholInfo;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
